# Templatize on the need for the integral of the third kind