--- a/benchmarks/elliptic_integrals_benchmark.cpp
+++ b/benchmarks/elliptic_integrals_benchmark.cpp
@@ -18,6 +18,35 @@ using quantities::si::Radian;
 
 namespace numerics {
 
+void BM_EllipticF(benchmark::State& state) {
+  constexpr int size = 20;
+
+  std::mt19937_64 random(42);
+  std::uniform_real_distribution<> distribution_φ(0.0, π / 2);
+  std::uniform_real_distribution<> distribution_n(0.0, 1.0);
+  std::uniform_real_distribution<> distribution_mc(0.0, 1.0);
+  std::vector<Angle> φs;
+  std::vector<double> ns;
+  std::vector<double> mcs;
+  for (int i = 0; i < size; ++i) {
+    φs.push_back(distribution_φ(random) * Radian);
+    ns.push_back(distribution_n(random));
+    mcs.push_back(distribution_mc(random));
+  }
+
+  while (state.KeepRunningBatch(size * size * size)) {
+    Angle f;
+    for (Angle const φ : φs) {
+      for (double const n : ns) {
+        for (double const mc : mcs) {
+          f += EllipticF(φ, mc);
+        }
+      }
+    }
+    benchmark::DoNotOptimize(f);
+  }
+}
+
 void BM_EllipticFEΠ(benchmark::State& state) {
   constexpr int size = 20;
 
@@ -84,6 +113,7 @@ void BM_FukushimaEllipticBDJ(benchmark::State& state) {
   }
 }
 
+BENCHMARK(BM_EllipticF);
 BENCHMARK(BM_EllipticFEΠ);
 BENCHMARK(BM_FukushimaEllipticBDJ);