This makes it possible to compute F and E without computing J when Π is not needed.

Before:
```
Run on (4 X 2808 MHz CPU s)
CPU Caches:
  L1 Data 32K (x2)
  L1 Instruction 32K (x2)
  L2 Unified 262K (x2)
  L3 Unified 4194K (x1)
------------------------------------------------------
Benchmark               Time           CPU Iterations
------------------------------------------------------
BM_EllipticF          233 ns        228 ns    2951600
BM_EllipticFEΠ        235 ns        230 ns    2992000
```
After:
```
Run on (4 X 2808 MHz CPU s)
CPU Caches:
  L1 Data 32K (x2)
  L1 Instruction 32K (x2)
  L2 Unified 262K (x2)
  L3 Unified 4194K (x1)
------------------------------------------------------
Benchmark               Time           CPU Iterations
------------------------------------------------------
BM_EllipticF          133 ns        123 ns    4072800
BM_EllipticFEΠ        239 ns        230 ns    2992000
```